requirements.txt
fastapi
uvicorn
fastai
torch
newspaper3k
